# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Notification:
    properties:
      created_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      resource_id:
        type: string
      user_id:
        type: string
    type: object
  errors.APIErrorResponse:
    properties:
      code:
        type: string
      fields:
        items:
          $ref: '#/definitions/errors.FieldError'
        type: array
      message:
        type: string
      meta: {}
      request_id:
        type: string
      status:
        type: integer
    type: object
  errors.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
...
  requests.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  requests.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  requests.UserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  requests.VerificationRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  responses.LoginResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/responses.UserResponse'
    type: object
  responses.PaginatedResponse:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  responses.SimpleMessage:
    properties:
      data: {}
      message:
        example: Action completed successfully
        type: string
    type: object
...
  responses.TokenResponse:
    properties:
      expires_in:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
    type: object
...
  responses.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a comprehensive API for creating and managing appointments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Asiko API
  version: "1.0"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      operationId: refreshToken
      parameters:
      - description: Refresh token
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TokenResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Could not generate token
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend a verification code for a pending user registration.
      operationId: resendVerification
      parameters:
      - description: Email to resend verification code to
        in: body
        name: resend
        required: true
        schema:
          $ref: '#/definitions/requests.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/responses.SimpleMessage'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "404":
          description: Pending registration not found
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "409":
          description: Account already verified
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Resend verification code
      tags:
      - Authentication
  /auth/verify-registration:
    post:
      consumes:
      - application/json
      description: Verify a user's email address with a code to complete registration.
      operationId: verifyRegistration
      parameters:
      - description: Email and Verification Code
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/requests.VerificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "400":
          description: Invalid request payload or verification error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Verify user registration
      tags:
      - Authentication
...
  /notifications:
    get:
      description: Retrieves a paginated list of notifications for the currently authenticated
        user.
      operationId: getNotifications
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.PaginatedResponse'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entities.Notification'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - Notifications
  /notifications/read-all:
    put:
      description: Marks all notifications for the currently authenticated user as
        read.
      operationId: markAllNotificationsAsRead
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SimpleMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user in the system. This will trigger an email verification.
      operationId: createUser
      parameters:
      - description: User Registration Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.UserRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/responses.SimpleMessage'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIErrorResponse'
      summary: Create a new user (initiate registration)
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer' followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"